// src/app/dashboard/Models/jobs-count.model.ts
export class dashboard {

  created: number = 0;
  totalPositions: number = 0;
  jobPosts: number = 0;
  criticalRoles: number = 0;
}

export class pipeline {
  candidatesApplied: number = 0;
  candidatesShortListed: number = 0;
  candidatesOffered: number = 0;
  candidatesRejected: number = 0;
  candidatesSourced: number = 0
 
}